requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chat-downloader==0.2.8
click==8.1.7
colorlog==6.9.0
docstring_parser==0.16
Flask==3.1.0
frozenlist==1.5.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
iso8601==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kick.py @ git+https://github.com/gub-7/kick.py
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
twitchio==2.10.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
yarl==1.17.1
